PLG-97: moving product options functionality to additional helper class
